refactor: [Service] 객체지향적인 코드를 위해 Service와 Impl을 분리
