+++ src/main/java/com/example/good_match/domain/board/domain/BoardStatus.java
@@ -1,8 +1,6 @@
 package com.example.good_match.domain.board.domain;
 
 public enum BoardStatus {
-    reservation_wait,
-    reservation_complete,
-    reservation_cancel,
-    game_over
+    recruiting,
+    recruitment_complete
 }
