+++ src/main/java/com/example/good_match/domain/main/controller/MainController.java
@@ -1,0 +1,2 @@
+package com.example.good_match.domain.main.controller;public class MainController {
+}
