+++ src/main/java/com/example/good_match/domain/main/dto/response/MainBoardDto.java
@@ -1,0 +1,2 @@
+package com.example.good_match.domain.main.dto.response;public class MainBoardDto {
+}
