+++ src/main/java/com/example/good_match/domain/main/dto/response/MainResponseDto.java
@@ -1,0 +1,2 @@
+package com.example.good_match.domain.main.dto.response;public class MainResponseDto {
+}
