+++ src/main/java/com/example/good_match/domain/main/service/MainService.java
@@ -1,0 +1,2 @@
+package com.example.good_match.domain.main.service;public class MainService {
+}
